// apps/web/src/payload-generated-schema.ts
export const enum_todo_status = pgEnum("enum_todo_status", [
  "PENDING",
  "STARTED",
  "DONE",
]);
// ...
export const todo = pgTable(
  "todo",
  {
    id: uuid("id").defaultRandom().primaryKey(),
    title: varchar("title").notNull(),
    description: varchar("description"),
    status: enum_todo_status("status").notNull().default("PENDING"),
    updatedAt: timestamp("updated_at", {
      mode: "string",
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
    createdAt: timestamp("created_at", {
      mode: "string",
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
  },
  (columns) => [
    index("todo_updated_at_idx").on(columns.updatedAt),
    index("todo_created_at_idx").on(columns.createdAt),
  ],
);